% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim Space at beginning and end}
\usage{
trim(string)
}
\arguments{
\item{string:}{the part you are going to pass on}
}
\description{
This function allow you to cut unwanted space at the beginning and end due to unknow format from extraction
}
\examples{
trim(" ab c   ")
}
\keyword{string}
